+++ dev/src/ts/states/preload.ts
@@ -1,8 +1,6 @@
 export default class PreloadState extends Lightning.State {
 
     create() {
-        let g = new Lightning.Graphics();
-
         this.game.states.start('menu');
     }
 }